update rule checking cli

+++ edm_rule_check_CLI.py
@@ -29,13 +29,13 @@ from rule_utils import get_rule_list, get_obj_list, get_rule_img, check_consiste
 os.environ['CUDA_VISIBLE_DEVICES']='0'
 device = torch.device('cpu')
 
-def batch_load_samples_infer_rules(samples_dir, epoch_list, encoding="onehot"):
+def batch_load_samples_infer_rules(samples_dir, epoch_list, encoding="onehot", fmt="tensor_%s.pt"):
     rules_all = []
     for epoch in tqdm(epoch_list): 
-        if not os.path.exists(join(samples_dir, 'tensor_'+str(epoch)+'.pt')):
+        if not os.path.exists(join(samples_dir, fmt % epoch)):
             print(epoch, "not exist")
             break
-        samples = torch.load(join(samples_dir, 'tensor_'+str(epoch)+'.pt')) # (batch, 27, 9, 9)
+        samples = torch.load(join(samples_dir, fmt % epoch)) # (batch, 27, 9, 9)
         if encoding == "onehot":
             attr_tsr_list = onehot2attr_tsr(samples, threshold=0.5)
         elif encoding == "digit":
@@ -95,6 +95,34 @@ def visualize_rule_validity(epoch_list, rules_all, title_str="Wide Dep x3 Blnr",
     return fig
 
 
+
+def rule_summary_table(rules_all, consistent_mat, epoch_list):
+    rule_summary_df = []
+    assert rules_all.shape[0] == consistent_mat.shape[0] == len(epoch_list)
+    for rule_ep, consistent_ep, epoch in zip(rules_all, consistent_mat, epoch_list): 
+        rule_valid = (rule_ep != -1).mean()
+        rule_valid_cnt = (rule_ep != -1).sum()
+        rule_consistent_1 = (consistent_ep == 1).mean()
+        rule_consistent_1_cnt = (consistent_ep == 1).sum()
+        rule_consistent_2 = (consistent_ep == 2).mean()
+        rule_consistent_2_cnt = (consistent_ep == 2).sum()
+        rule_consistent_3 = (consistent_ep == 3).mean()
+        rule_consistent_3_cnt = (consistent_ep == 3).sum()
+        rule_summary_df.append({"epoch": epoch,
+                                "valid": rule_valid.mean(),
+                                "valid_cnt": rule_valid_cnt,
+                                "cst_1": rule_consistent_1.mean(),
+                                "cst_1_cnt": rule_consistent_1_cnt,
+                                "cst_2": rule_consistent_2.mean(),
+                                "cst_2_cnt": rule_consistent_2_cnt,
+                                "cst_3": rule_consistent_3.mean(),
+                                "cst_3_cnt": rule_consistent_3_cnt,
+                                })
+    rule_summary_df = pd.DataFrame(rule_summary_df)
+    print(rule_summary_df.tail())
+    return rule_summary_df
+
+
 if __name__ == "__main__":  
     import argparse
     parser = argparse.ArgumentParser(description="Process and visualize EDM rules.")
@@ -110,11 +138,12 @@ if __name__ == "__main__":
     parser.add_argument("--figname", type=str, default="", help="Name of the figure to save.")
     parser.add_argument("--figdir", type=str, default="Figures", help="Directory to save the figure.")
     parser.add_argument("--title_str", type=str, default="", help="Title of the figure.")
+    parser.add_argument("--fmt", type=str, default="tensor_%s.pt", help="Format of the sample files.")
+    parser.add_argument("--update", action="store_true", help="Update the existing inferred rules.")
                         
 
     # Parse arguments
     args = parser.parse_args()
-
     # exproot = "/n/holylabs/LABS/kempner_fellows/Users/binxuwang/DL_Projects/mini_edm/exps"
     # expname = "WideBlnr_RAVEN10_abstract_onehot_20240211-1743"
     # epoch_list = list(np.arange(0, 1000000, 5000))
@@ -127,16 +156,41 @@ if __name__ == "__main__":
     title_str = expname.replace("_"," ") if args.title_str == "" else args.title_str
     
     expdir = join(exproot, expname)
-    df = parse_train_logfile(join(expdir, "std.log"))
-    df.plot(x="step", y=["average_loss","batch_loss"], alpha=0.75)
-    plt.savefig(join(expdir, "loss_curve.png"))
-
-    rules_all_wide, consistent_mat_wide, epoch_list = batch_load_samples_infer_rules(
-        join(expdir, "samples"), epoch_list, encoding=encoding)
-    np.savez(join(expdir,"samples_inferred_rule_consistency.npz"), 
+    print("Analyzing", expname)
+    if os.path.exists(join(expdir, "std.log")):
+        df = parse_train_logfile(join(expdir, "std.log"))
+        df.plot(x="step", y=["average_loss","batch_loss"], alpha=0.75)
+        plt.savefig(join(expdir, "loss_curve.png"))
+    
+    if os.path.exists(join(expdir, "samples_inferred_rule_consistency.npz")) and args.update:
+        print("Inferred rules already exist. Loading from file.")
+        npzfile = np.load(join(expdir, "samples_inferred_rule_consistency.npz"))
+        rules_all_load = npzfile["rules_all"]
+        consistent_mat_load = npzfile["consistent_mat"]
+        epoch_list_load = npzfile["epoch_list"]
+        epoch_rest = [epoch for epoch in epoch_list if epoch not in epoch_list_load]
+        print(rules_all_load.shape, consistent_mat_load.shape, len(epoch_rest))
+        print("Updating inferred rules, starting from epoch: ", epoch_rest[0])
+        rules_all_rest, consistent_mat_rest, epoch_rest = batch_load_samples_infer_rules(
+            join(expdir, "samples"), epoch_rest, encoding=encoding, fmt=args.fmt)
+        if len(epoch_rest) == 0 and len(rules_all_rest) == 0:
+            rules_all_wide, consistent_mat_wide, epoch_list = rules_all_load, consistent_mat_load, epoch_list_load
+        else:
+            rules_all_wide = np.concatenate([rules_all_load, rules_all_rest], axis=0)
+            consistent_mat_wide = np.concatenate([consistent_mat_load, consistent_mat_rest], axis=0)
+            epoch_list = np.concatenate([epoch_list_load, epoch_rest], axis=0)
+            print("Inferred rules updated.")
+        print(rules_all_wide.shape, consistent_mat_wide.shape, len(epoch_list))
+    else:
+        rules_all_wide, consistent_mat_wide, epoch_list = batch_load_samples_infer_rules(
+            join(expdir, "samples"), epoch_list, encoding=encoding, fmt=args.fmt)
+    
+    np.savez(join(expdir, "samples_inferred_rule_consistency.npz"), 
             consistent_mat=consistent_mat_wide, 
             rules_all=rules_all_wide, epoch_list=epoch_list)
-
+    
+    rule_summary_df = rule_summary_table(rules_all_wide, consistent_mat_wide, epoch_list)
+    
     visualize_consistency(epoch_list, consistent_mat_wide, 
                         title_str=title_str, figname=figname, figdir=figdir,)
